+++ external/skia/src/core/SkVarAlloc.cpp
@@ -7,6 +7,8 @@
 
 #include "SkVarAlloc.h"
 
+#include "SkMalloc.h"
+
 struct SkVarAlloc::Block {
     Block* prev;
     char* data() { return (char*)(this + 1); }
@@ -45,7 +47,7 @@ SkVarAlloc::~SkVarAlloc() {
 void SkVarAlloc::makeSpace(size_t bytes) {
     SkASSERT(SkIsAlignPtr(bytes));
 
-    size_t alloc = 1<<fLgSize++;
+    size_t alloc = static_cast<size_t>(1)<<fLgSize++;
     while (alloc < bytes + sizeof(Block)) {
         alloc *= 2;
     }
