+++ external/xmlrpcpp/src/XmlRpcServerConnection.cpp
@@ -347,7 +347,7 @@ XmlRpcServerConnection::generateHeader(std::string const& body)
     "Content-length: ";
 
   char buffLen[40];
-  sprintf(buffLen,"%d\r\n\r\n", body.size());
+  sprintf(buffLen,"%d\r\n\r\n", (int)body.size());
 
   return header + buffLen;
 }
